+++ tmall-parent/service/service-category/src/main/java/com/study/tmall/category/controller/PropertyController.java
@@ -33,10 +33,10 @@ public class PropertyController {
     @GetMapping("/list/{current}/{limit}")
     public Result findPageProperty(
             @ApiParam(name = "current", value = "起始页", required = true)
-            @PathVariable Integer current,
+            @PathVariable Long current,
 
             @ApiParam(name = "limit", value = "每页记录数", required = true)
-            @PathVariable Integer limit,
+            @PathVariable Long limit,
 
             @ApiParam(name = "propertyQueryVo", value = "条件查询vo")
             PropertyQueryVo propertyQueryVo){
