+++ tmall-parent/service/service-order/src/main/java/com/study/tmall/order/controller/OrderInfoController.java
@@ -32,10 +32,10 @@ public class OrderInfoController {
     @GetMapping("/list/{current}/{limit}")
     public Result findPageOrderInfo(
             @ApiParam(name = "current", value = "起始页", required = true)
-            @PathVariable Integer current,
+            @PathVariable Long current,
 
             @ApiParam(name = "limit", value = "每页记录", required = true)
-            @PathVariable Integer limit,
+            @PathVariable Long limit,
 
             @ApiParam(name = "orderQueryVo", value = "查询条件vo")
             OrderQueryVo orderQueryVo){
