+++ tmall-parent/service/service-product/src/main/java/com/study/tmall/product/controller/ProductInfoController.java
@@ -35,10 +35,10 @@ public class ProductInfoController {
     @GetMapping("/list/{current}/{limit}")
     public Result findPageProductInfo(
             @ApiParam(name = "current", value = "起始页", required = true)
-            @PathVariable Integer current,
+            @PathVariable Long current,
 
             @ApiParam(name = "limit", value = "每页记录数", required = true)
-            @PathVariable Integer limit,
+            @PathVariable Long limit,
 
             @ApiParam(name = "productQueryVo", value = "条件查询商品vo")
             ProductQueryVo productQueryVo){
